Add cookie recipe page to Tutorial 3 challenge
- Refactor Tutorial 3 challenge to have abstract Tutorial3Recipe, which both the brownie and cookie recipes extend.
- CookieRecipe overrides the success checker and callback so that when the user clicks the checkbox referring to an egg ingredient, they complete the module.

Only thing missing now is the actual instructions. I'm *still* not sure how we're meant to be doing these, as the instructions for the challenges can be very lengthy (and therefore take up a lot of space on the screen if we show them as text). Might look into using semantics/GestureDectector to make a button that says the instructions when the user presses it.